+++ packages/studio/src/hooks/gsapScriptCommitTypes.ts
@@ -1,9 +1,6 @@
 import type { ParsedGsap } from "@hyperframes/core/gsap-parser";
-import type { Composition } from "@hyperframes/sdk";
 import type { DomEditSelection } from "../components/editor/domEditingTypes";
 import type { EditHistoryKind } from "../utils/editHistory";
-import type { ShadowGsapOp } from "../utils/sdkShadow";
-import type { ShadowKeyframeOp } from "../utils/sdkShadowGsapKeyframe";
 
 export interface MutationResult {
   ok: boolean;
@@ -28,10 +25,6 @@ export interface CommitMutationOptions {
    * (and under distinct keys) run concurrently as before.
    */
   serializeKey?: string;
-  /** Stage 7 Step 3b: typed SDK equivalent of this mutation for value-fidelity shadow. */
-  shadowGsapOp?: ShadowGsapOp;
-  /** Typed SDK equivalent of a keyframe mutation for keyframe value-fidelity shadow (gsap_keyframe). */
-  shadowKeyframeOp?: ShadowKeyframeOp;
 }
 
 export type CommitMutation = (
@@ -70,6 +63,4 @@ export interface GsapScriptCommitsParams {
   onCacheInvalidate: () => void;
   onFileContentChanged?: (path: string, content: string) => void;
   showToast: (message: string, tone?: "error" | "info") => void;
-  /** Stage 7 Step 3b: SDK session for shadow GSAP dispatch (server stays authoritative). */
-  sdkSession?: Composition | null;
 }
