+++ packages/studio/src/hooks/useTimelineEditing.ts
@@ -1,11 +1,7 @@
 // Pre-existing-complex timeline hook (DOM patch + GSAP position shift/scale +
-// playback-start resolution); this PR adds guarded shadow-timing dispatches in
-// the move/resize .then() chains, which nudges several callbacks over the CC
-// threshold. The added branches are telemetry-only.
+// playback-start resolution).
 // fallow-ignore-file complexity
 import { useCallback, useRef } from "react";
-import type { Composition } from "@hyperframes/sdk";
-import { runShadowDelete, runShadowTiming } from "../utils/sdkShadow";
 import type { TimelineElement } from "../player";
 import { usePlayerStore } from "../player";
 import { useRazorSplit } from "./useRazorSplit";
@@ -60,8 +56,6 @@ interface UseTimelineEditingOptions {
   pendingTimelineEditPathRef: React.MutableRefObject<Set<string>>;
   uploadProjectFiles: (files: Iterable<File>, dir?: string) => Promise<string[]>;
   isRecordingRef?: React.RefObject<boolean>;
-  /** Stage 7 Step 3b: SDK session for shadow timing dispatch (server stays authoritative). */
-  sdkSession?: Composition | null;
 }
 
 // ── Hook ──
@@ -79,7 +73,6 @@ export function useTimelineEditing({
   pendingTimelineEditPathRef,
   uploadProjectFiles,
   isRecordingRef,
-  sdkSession,
 }: UseTimelineEditingOptions) {
   const projectIdRef = useRef(projectId);
   projectIdRef.current = projectId;
@@ -148,11 +141,6 @@ export function useTimelineEditing({
           value: String(updates.track),
         });
       }).then(() => {
-        if (sdkSession)
-          runShadowTiming(sdkSession, element.hfId, {
-            start: updates.start,
-            trackIndex: updates.track,
-          });
         const pid = projectIdRef.current;
         if (delta !== 0 && element.domId && pid) {
           return shiftGsapPositions(pid, filePath, element.domId, delta)
@@ -161,7 +149,7 @@ export function useTimelineEditing({
         }
       });
     },
-    [previewIframeRef, enqueueEdit, activeCompPath, reloadPreview, sdkSession],
+    [previewIframeRef, enqueueEdit, activeCompPath, reloadPreview],
   );
 
   const handleTimelineElementResize = useCallback(
@@ -205,11 +193,6 @@ export function useTimelineEditing({
         }
         return patched;
       }).then(() => {
-        if (sdkSession)
-          runShadowTiming(sdkSession, element.hfId, {
-            start: updates.start,
-            duration: updates.duration,
-          });
         const pid = projectIdRef.current;
         if (timingChanged && element.domId && pid) {
           return scaleGsapPositions(
@@ -227,7 +210,7 @@ export function useTimelineEditing({
         return reloadPreview();
       });
     },
-    [previewIframeRef, enqueueEdit, activeCompPath, reloadPreview, sdkSession],
+    [previewIframeRef, enqueueEdit, activeCompPath, reloadPreview],
   );
 
   const handleTimelineElementDelete = useCallback(
@@ -288,7 +271,6 @@ export function useTimelineEditing({
           );
         usePlayerStore.getState().setSelectedElementId(null);
         reloadPreview();
-        if (sdkSession) runShadowDelete(sdkSession, element.hfId);
         showToast(`Deleted ${label}. Use Undo to restore it.`, "info");
       } catch (error) {
         const message = error instanceof Error ? error.message : "Failed to delete timeline clip";
@@ -304,7 +286,6 @@ export function useTimelineEditing({
       domEditSaveTimestampRef,
       reloadPreview,
       isRecordingRef,
-      sdkSession,
     ],
   );
