+++ packages/studio/src/utils/sdkCutover.ts
@@ -1,10 +1,9 @@
 import type { MutableRefObject } from "react";
-import type { Composition } from "@hyperframes/sdk";
+import type { Composition, EditOp } from "@hyperframes/sdk";
 import type { DomEditSelection } from "../components/editor/domEditing";
 import type { EditHistoryKind } from "./editHistory";
 import type { PatchOperation } from "./sourcePatcher";
 import { STUDIO_SDK_CUTOVER_ENABLED } from "../components/editor/manualEditingAvailability";
-import { patchOpsToSdkEditOps } from "./sdkShadow";
 import { trackStudioEvent } from "./studioTelemetry";
 
 const CUTOVER_OP_TYPES = new Set<PatchOperation["type"]>([
@@ -13,6 +12,42 @@ const CUTOVER_OP_TYPES = new Set<PatchOperation["type"]>([
   "attribute",
   "html-attribute",
 ]);
+
+/**
+ * Map Studio PatchOperations for a given hf-id to SDK EditOps.
+ *
+ * Multiple inline-style ops are coalesced into a single setStyle (SDK batches
+ * style changes naturally). One SDK op is emitted per non-style op.
+ */
+function patchOpsToSdkEditOps(hfId: string, ops: PatchOperation[]): EditOp[] {
+  const result: EditOp[] = [];
+  const styles: Record<string, string | null> = {};
+  let hasStyles = false;
+
+  for (const op of ops) {
+    if (op.type === "inline-style") {
+      styles[op.property] = op.value;
+      hasStyles = true;
+    } else if (op.type === "text-content") {
+      result.push({ type: "setText", target: hfId, value: op.value ?? "" });
+    } else if (op.type === "attribute") {
+      result.push({
+        type: "setAttribute",
+        target: hfId,
+        name: op.property.startsWith("data-") ? op.property : `data-${op.property}`,
+        value: op.value,
+      });
+    } else if (op.type === "html-attribute") {
+      result.push({ type: "setAttribute", target: hfId, name: op.property, value: op.value });
+    }
+  }
+
+  if (hasStyles) {
+    result.unshift({ type: "setStyle", target: hfId, styles });
+  }
+
+  return result;
+}
 
 export function shouldUseSdkCutover(
   flagEnabled: boolean,
